pull github stuff into its own class
